Update checkout.php
changed the CSS for the page so that it looks cleaner, and added parameters to the details the user enters making sure that they have to enter 16 digits for the card number. also added a function that checks that they select international delivery if their address is outside the UK, and made them enter the details in a multi-step form, the pricing does change based on what shipping method they select.